+++ refinedstorage-jei-integration-common/src/main/java/com/refinedmods/refinedstorage/jei/common/Platform.java
@@ -1,9 +1,12 @@
 package com.refinedmods.refinedstorage.jei.common;
 
+import com.refinedmods.refinedstorage.api.resource.ResourceAmount;
 import com.refinedmods.refinedstorage.common.support.resource.FluidResource;
 
 import java.util.Optional;
 
 public interface Platform {
     Optional<FluidResource> convertJeiIngredientToFluid(Object ingredient);
+
+    Optional<ResourceAmount> convertJeiIngredientToFluidAmount(Object ingredient);
 }